feat: add word filters and blacklist (#98)
This update introduce two config on chat_list.json

filters (Optional) - An array of strings to filter words. If the message containes any of the strings in the array, it WILL BE forwarded.

blacklist (Optional) - An array of strings to blacklist words. If the message containes any of the string in the array, it will NOT BE forwarded.